% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FantasyEconomics.R
\name{FantasyEconomics}
\alias{FantasyEconomics}
\alias{fantasyEconomicsUI}
\alias{fantasyEconomics}
\title{Fantasy Economics}
\usage{
fantasyEconomicsUI()

fantasyEconomics(input, output, session)
}
\arguments{
\item{input}{Shiny Input Parameter}

\item{output}{Shiny Output Parameter}

\item{session}{Shiny Session Parameter}
}
\value{
Shiny ui and server code to run the shiny application.
}
\description{
An economics game used for inductions by the Government Economic Service
that simulates the effects of policies on key macroeconomic measures of
national output and inequality.
}
\details{
The shiny application is comprised of two modules:
\itemize{
  \item{\code{setup}}{A module to configure the application}
  \item{\code{game}}{A module that contains the game code}
}
}
\section{Functions}{
\itemize{
\item \code{fantasyEconomicsUI}: Fantasy Economics Shiny User Interface

\item \code{fantasyEconomics}: Fantasy Economics Shiny Server Code
}}

