% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game_module}
\alias{game_module}
\alias{gameUI}
\alias{game}
\title{Game Module}
\usage{
gameUI(id)

game(input, output, session, setup)
}
\arguments{
\item{id}{Namespace value}

\item{input}{Shiny input parameter}

\item{output}{Shiny output parameter}

\item{session}{Shiny session parameter}

\item{setup}{Reactive values from the setup module}
}
\value{
Reactive Values list:
\itemize{
  \item{\code{game_round}}{The current gameround}
  \item{\code{submit}}{Boolean value on whether or not to submit data}
  \item{\code{data_in}}{The data contained in the rhandsontable}
  \item{\code{previous_data}}{saved data from previous round}
}
}
\description{
Game Module
}
\details{
The game module contains the following outputs:
\itemize{
  \item{\code{shock}}{A description of the economic shock that has occured in that round}
  \item{\code{data_table}}{The data table for inputting policy choices and viewing effects}
  \item{\code{gdp_chart}}{Chart to tracks the gdp score by round}
  \item{\code{gini_chart}}{Chart to track gini score by round}
}
}
\section{Functions}{
\itemize{
\item \code{gameUI}: User Interface for the Game Module

\item \code{game}: Shiny Server Code for the Game Module
}}

